+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.7.3
+* Bugfix
+  - fix: `props.scrollRowByRow` support param: `scrollbar`. If scroll horizontal scrollbar should not trigger scrollRowByRow
+
 ## 0.7.2
 * Feature
   - `props.scrollRowByRow` support param: `scrollbar`. Dealing the problem: if we drag scrollbar very fast ,and the table container show empty area shortly.